backend/requirements.txt
uvicorn
fastapi
pydantic
pandas
logging
requests
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud
google-cloud-aiplatform
transformers
tensorflow